adduser: refactor legacy login into its own module
PR-URL: https://github.com/npm/npm/pull/13389
Credit: @zkat
